add comment function

+++ src/main/java/com/lau/community/service/QuestionService.java
@@ -4,8 +4,6 @@ import com.lau.community.dto.PaginationDTO;
 import com.lau.community.dto.QuestionDTO;
 import com.lau.community.model.Question;
 
-import java.util.List;
-
 /**
  * @author ：lauchun
  * @date ：Created in 2020-02-25
@@ -17,9 +15,9 @@ public interface QuestionService {
 
     PaginationDTO list(Integer id, Integer page, Integer size);
 
-    QuestionDTO findById(Integer id);
+    QuestionDTO findById(Long id);
 
     void createOrUpdate(Question question);
 
-    void incView(Integer id);
+    void incView(Long id);
 }
